TheBotFamily/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
requests
pydantic
supabase
gspread
oauth2client
